requirements.txt
alabaster==0.7.12
Babel==2.10.3
cachelib==0.6.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
docutils==0.17.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Session==0.4.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.3
pefile==2021.5.24
py-algorand-sdk==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.11.0
Pygments==2.12.0
PyNaCl==1.4.0
pyparsing==3.0.9
pyteal==0.10.1
pytz==2022.1
requests==2.27.1
rsa==4.7.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
Telethon==1.24.0
urllib3==1.26.8
Werkzeug==2.0.2
